dht_node: use externally mapped ports
The upnp gateway may map an external port that is different
form the one set into the configuration object. For this reason
it is important that the various components are notified about
the external port value.

Pass the external_{dht_node,peer}_port value when instantiating
the dht_node object.

Signed-off-by: Antonio Quartulli <antonio@mandelbit.com>