conf: allow user to manually set external IP
When testing or when running lbrynet in a controlled
environment, it may make sense to manually set the external
IP of the node instead of letting it autodetect its address.

Add a new key 'external_ip' in the config file for this purpose.
When set, the node will skip the autodetection.

Signed-off-by: Antonio Quartulli <antonio@mandelbit.com>